// src/types/nodeType.ts
enum NodeType {
	Document = 'DOCUMENT',
	Canvas = 'CANVAS',
	Frame = 'FRAME',
	Group = 'GROUP',
	Section = 'SECTION',
	Vector = 'VECTOR',
	BooleanOperation = 'BOOLEAN_OPERATION',
	Star = 'STAR',
	Line = 'LINE',
	Ellipse = 'ELLIPSE',
	RegularPolygon = 'REGULAR_POLYGON',
	Rectangle = 'RECTANGLE',
	Table = 'TABLE',
	TableCell = 'TABLE_CELL',
	Text = 'TEXT',
	Slice = 'SLICE',
	Component = 'COMPONENT',
	ComponentSet = 'COMPONENT_SET',
	Instance = 'INSTANCE',
	Sticky = 'STICKY',
	ShapeWithText = 'SHAPE_WITH_TEXT',
	Connector = 'CONNECTOR',
	WashiTape = 'WASHI_TAPE'
}

export default NodeType;